adding variables as unary constraints

+++ src/Tools/Variable.java
@@ -3,7 +3,7 @@ package Tools;
 public class Variable
 {
 	public String name;
-	public float confidence;
+	public double confidence;
 	public int domain;
 
 	public Variable()
